Linked Map marker to info
The markers on the map now act as buttons, so when they are clicked the info pops up on the right panel. Created a separate JS file to have these function, simply for organization's sake. The code is very sloppy and could have been done in a much neater and more efficient way, BUT it functions correctly.